Fix compiler warning about comparison of signed and unsigned values.
--HG--
branch : ctypes
extra : convert_revision : svn%3A6015fed2-1504-0410-9fe1-9d1591cc4771/ctypes/trunk/ctypes%4050304

+++ source/callproc.c
@@ -1459,7 +1459,7 @@ resize(PyObject *self, PyObject *args)
 				"excepted ctypes instance");
 		return NULL;
 	}
-	if (size < (size_t)dict->size) {
+	if (size < dict->size) {
 		PyErr_Format(PyExc_ValueError,
 			     "minimum size is %d", dict->size);
 		return NULL;
